move examples to recipes
