+++ docs/source/noether/recipe_zoo.rst
@@ -1,0 +1,23 @@
+Noether Recipe Zoo
+==================
+
+In the `Noether Recipe Zoo <https://github.com/Emmi-AI/noether/tree/main/recipes>`_ we currently provide the following recipes:
+
+.. list-table::
+   :header-rows: 1
+   :widths: 30 35 35
+
+   * - Recipe
+     - Source Code
+     - Link
+   * - **External Aerodynamics**
+     - `recipes/aero_cfd/ <https://github.com/Emmi-AI/noether/tree/main/recipes/aero_cfd>`_
+     - :doc:`/tutorials/walkthrough/index`
+   * - **External Aerodynamics (Python)**
+     - `recipes/aero_cfd_python/ <https://github.com/Emmi-AI/noether/tree/main/recipes/aero_cfd_python>`_
+     - :doc:`aero_cfd_python`
+
+.. toctree::
+   :hidden:
+
+   aero_cfd_python
