--- recipes/aero_cfd/jobs/__init__.py
@@ -1,2 +1,0 @@
-#  Copyright © 2025 Emmi AI GmbH. All rights reserved.
-
