+++ recipes/aero_cfd/jobs/train_drivaerml.job
@@ -9,6 +9,6 @@
 #SBATCH --nice=0
 #SBATCH --array=1-20%10
 
-EXPERIMENT_FILE=recipes/aero_cfd/jobs/experiments/drivaerml_experiments.txt
+EXPERIMENT_FILE=$(dirname "$0")/experiments/drivaerml_experiments.txt
 # run
 srun --kill-on-bad-exit=1 --cpus-per-task=28 uv run noether-train --hp recipes/aero_cfd/configs/train_drivaerml.yaml $(head -$SLURM_ARRAY_TASK_ID $EXPERIMENT_FILE | tail -1)
