+++ recipes/aero_cfd_python/presets/drivaernet.py
@@ -2,15 +2,18 @@
 
 from __future__ import annotations
 
-from examples.aero_cfd.presets.base import AeroCFDPreset, AeroPipelineParams
-from noether.core.schemas.dataset import AeroDataSpecs
+from typing import Any
+
+from noether.core.schemas.dataset import AeroDataSpecs, DatasetBaseConfig, DatasetWrappers
 from noether.core.schemas.normalizers import FieldNormalizerConfig
 
+from .base import AeroCFDPreset, AeroPipelineParams
 
-class EmmiWingPreset(AeroCFDPreset):
-    """Preset for the EMMI Wing CFD dataset."""
 
-    dataset_kind = "noether.data.datasets.cfd.EmmiWingDataset"
+class DrivAerNetPreset(AeroCFDPreset):
+    """Preset for the DrivAerNet++ CFD dataset."""
+
+    dataset_kind = "noether.data.datasets.cfd.DrivAerNetDataset"
 
     pipeline_defaults: AeroPipelineParams = {
         "num_surface_points": 16384,
@@ -37,41 +40,12 @@ class EmmiWingPreset(AeroCFDPreset):
         },
     }
 
-    forward_properties_map: dict[str, list[str]] = {
-        "noether.modeling.models.aerodynamics.AeroUPT": [
-            "surface_position_batch_idx",
-            "surface_position_supernode_idx",
-            "surface_position",
-            "surface_query_position",
-            "volume_query_position",
-        ],
-        "noether.modeling.models.aerodynamics.AeroABUPT": [
-            "geometry_position",
-            "geometry_supernode_idx",
-            "geometry_batch_idx",
-            "surface_anchor_position",
-            "volume_anchor_position",
-            "geometry_design_parameters",
-            "inflow_design_parameters",
-        ],
-        "_default": [
-            "surface_position",
-            "volume_position",
-            "surface_features",
-            "volume_features",
-        ],
-    }
-
     @property
     def data_specs(self) -> AeroDataSpecs:
         return AeroDataSpecs(
             position_dim=3,
             surface_output_dims={"pressure": 1, "friction": 3},
             volume_output_dims={"pressure": 1, "velocity": 3, "vorticity": 3},
-            conditioning_dims={
-                "geometry_design_parameters": 5,
-                "inflow_design_parameters": 2,
-            },
         )
 
     @property
@@ -81,20 +55,51 @@ class EmmiWingPreset(AeroCFDPreset):
             "surface_friction": FieldNormalizerConfig(strategy="mean_std"),
             "volume_pressure": FieldNormalizerConfig(strategy="mean_std"),
             "volume_velocity": FieldNormalizerConfig(strategy="mean_std"),
-            # Wing uses magnitude-based normalization for vorticity (mean=0, std=magnitude_mean)
             "volume_vorticity": FieldNormalizerConfig(
                 strategy="mean_std",
-                stat_keys={"mean": "_zero", "std": "volume_vorticity_magnitude_mean"},
+                logscale=True,
+                stat_keys={"mean": "volume_vorticity_logscale_mean", "std": "volume_vorticity_logscale_std"},
             ),
-            "geometry_design_parameters": FieldNormalizerConfig(strategy="mean_std"),
-            "inflow_design_parameters": FieldNormalizerConfig(strategy="mean_std"),
-            "surface_position": FieldNormalizerConfig(strategy="position", scale=1000),
-            "volume_position": FieldNormalizerConfig(strategy="position", scale=1000),
+            "surface_position": FieldNormalizerConfig(
+                strategy="position", scale=1000, stat_keys={"min": "raw_pos_min", "max": "raw_pos_max"}
+            ),
+            "volume_position": FieldNormalizerConfig(
+                strategy="position", scale=1000, stat_keys={"min": "raw_pos_min", "max": "raw_pos_max"}
+            ),
         }
 
     @property
     def excluded_properties(self) -> set[str]:
-        return {"surface_normals", "volume_normals", "volume_sdf"}
+        return {"surface_normals", "volume_normals", "volume_sdf", "surface_sdf"}
+
+    def build_dataset(
+        self,
+        *,
+        split: str,
+        root: str,
+        model_kind: str,
+        wrappers: list[DatasetWrappers] | None = None,
+        filter_categories: tuple[str, ...] | None = None,
+        **overrides: Any,
+    ) -> DatasetBaseConfig:
+        """Build dataset config with optional category filtering.
+
+        Args:
+            filter_categories: optional tuple of DrivAerNet design categories to include
+                (e.g., ``("F_S_WWS_WM", "N_S_WWS_WM")``). None loads all categories.
+        """
+        from noether.core.schemas.aero import AeroDatasetConfig
+
+        return AeroDatasetConfig(
+            kind=self.dataset_kind,
+            root=root,
+            split=split,
+            pipeline=self.build_pipeline(model_kind, **overrides),
+            dataset_normalizers=self.build_normalizers(),
+            dataset_wrappers=wrappers,
+            excluded_properties=self.excluded_properties,
+            filter_categories=filter_categories,
+        )
 
     def target_properties(self) -> list[str]:
         return [
