+++ recipes/aero_cfd_python/presets/emmi_wing.py
@@ -2,15 +2,16 @@
 
 from __future__ import annotations
 
-from examples.aero_cfd.presets.base import AeroCFDPreset, AeroPipelineParams
 from noether.core.schemas.dataset import AeroDataSpecs
 from noether.core.schemas.normalizers import FieldNormalizerConfig
 
+from .base import AeroCFDPreset, AeroPipelineParams
 
-class DrivAerMLPreset(AeroCFDPreset):
-    """Preset for the DrivAerML CFD dataset (CAEML benchmark)."""
 
-    dataset_kind = "noether.data.datasets.cfd.DrivAerMLDataset"
+class EmmiWingPreset(AeroCFDPreset):
+    """Preset for the EMMI Wing CFD dataset."""
+
+    dataset_kind = "noether.data.datasets.cfd.EmmiWingDataset"
 
     pipeline_defaults: AeroPipelineParams = {
         "num_surface_points": 16384,
@@ -37,12 +38,41 @@ class DrivAerMLPreset(AeroCFDPreset):
         },
     }
 
+    forward_properties_map: dict[str, list[str]] = {
+        "noether.modeling.models.aerodynamics.AeroUPT": [
+            "surface_position_batch_idx",
+            "surface_position_supernode_idx",
+            "surface_position",
+            "surface_query_position",
+            "volume_query_position",
+        ],
+        "noether.modeling.models.aerodynamics.AeroABUPT": [
+            "geometry_position",
+            "geometry_supernode_idx",
+            "geometry_batch_idx",
+            "surface_anchor_position",
+            "volume_anchor_position",
+            "geometry_design_parameters",
+            "inflow_design_parameters",
+        ],
+        "_default": [
+            "surface_position",
+            "volume_position",
+            "surface_features",
+            "volume_features",
+        ],
+    }
+
     @property
     def data_specs(self) -> AeroDataSpecs:
         return AeroDataSpecs(
             position_dim=3,
             surface_output_dims={"pressure": 1, "friction": 3},
             volume_output_dims={"pressure": 1, "velocity": 3, "vorticity": 3},
+            conditioning_dims={
+                "geometry_design_parameters": 5,
+                "inflow_design_parameters": 2,
+            },
         )
 
     @property
@@ -52,18 +82,24 @@ class DrivAerMLPreset(AeroCFDPreset):
             "surface_friction": FieldNormalizerConfig(strategy="mean_std"),
             "volume_pressure": FieldNormalizerConfig(strategy="mean_std"),
             "volume_velocity": FieldNormalizerConfig(strategy="mean_std"),
+            # Wing uses magnitude-based normalization for vorticity (mean=0, std=magnitude_mean)
             "volume_vorticity": FieldNormalizerConfig(
                 strategy="mean_std",
-                logscale=True,
-                stat_keys={"mean": "volume_vorticity_logscale_mean", "std": "volume_vorticity_logscale_std"},
+                stat_keys={"mean": "_zero", "std": "volume_vorticity_magnitude_mean"},
             ),
-            "surface_position": FieldNormalizerConfig(strategy="position", scale=1000),
-            "volume_position": FieldNormalizerConfig(strategy="position", scale=1000),
+            "geometry_design_parameters": FieldNormalizerConfig(strategy="mean_std"),
+            "inflow_design_parameters": FieldNormalizerConfig(strategy="mean_std"),
+            "surface_position": FieldNormalizerConfig(
+                strategy="position", scale=1000, stat_keys={"min": "raw_pos_min", "max": "raw_pos_max"}
+            ),
+            "volume_position": FieldNormalizerConfig(
+                strategy="position", scale=1000, stat_keys={"min": "raw_pos_min", "max": "raw_pos_max"}
+            ),
         }
 
     @property
     def excluded_properties(self) -> set[str]:
-        return {"surface_normals", "volume_normals", "volume_sdf"}
+        return {"surface_normals", "volume_normals", "volume_sdf", "surface_sdf"}
 
     def target_properties(self) -> list[str]:
         return [
