+++ recipes/aero_cfd_python/train_ahmedml.py
@@ -1,12 +1,20 @@
 #  Copyright © 2026 Emmi AI GmbH. All rights reserved.
 
-from examples.aero_cfd import ShapeNetCarPreset
+from noether.core.distributed.utils import accelerator_to_device
 from noether.training.runners import HydraRunner
 
-DATASET_ROOT = "/path/to/shapenet_car"
-OUTPUT_PATH = "/path/to/outputs/shapenet_car"
+from .presets import AhmedMLPreset
+
+DATASET_ROOT = "/path/to/ahmedml"
+OUTPUT_PATH = "/path/to/outputs/ahmedml"
 TRAINER_KIND = "noether.training.trainers.WeightedLossTrainer"
-FIELD_WEIGHTS = {"surface_pressure": 1.0, "volume_velocity": 1.0}
+FIELD_WEIGHTS = {
+    "surface_pressure": 1.0,
+    "surface_friction": 1.0,
+    "volume_pressure": 1.0,
+    "volume_velocity": 1.0,
+    "volume_vorticity": 1.0,
+}
 
 
 def train_abupt(
@@ -14,10 +22,9 @@ def train_abupt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains AB-UPT model using ShapeNetCar dataset."""
-    preset = ShapeNetCarPreset()
+    """Trains AB-UPT model using AhmedML dataset."""
+    preset = AhmedMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroABUPT",
         model_params=dict(hidden_dim=192, geometry_depth=6, physics_blocks=["perceiver"] + ["shared", "cross"] * 5),
@@ -25,10 +32,11 @@ def train_abupt(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
+        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_upt(
@@ -36,10 +44,9 @@ def train_upt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains UPT model using ShapeNetCar dataset."""
-    preset = ShapeNetCarPreset()
+    """Trains UPT model using AhmedML dataset."""
+    preset = AhmedMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroUPT",
         model_params=dict(hidden_dim=192, num_heads=3, approximator_depth=12),
@@ -47,10 +54,11 @@ def train_upt(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
+        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transformer(
@@ -58,10 +66,9 @@ def train_transformer(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains Transformer model using ShapeNetCar dataset."""
-    preset = ShapeNetCarPreset()
+    """Trains Transformer model using AhmedML dataset."""
+    preset = AhmedMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransformer",
         model_params=dict(hidden_dim=192, depth=12),
@@ -69,10 +76,11 @@ def train_transformer(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
+        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transolver(
@@ -80,10 +88,9 @@ def train_transolver(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains Transolver model using ShapeNetCar dataset."""
-    preset = ShapeNetCarPreset()
+    """Trains Transolver model using AhmedML dataset."""
+    preset = AhmedMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransolver",
         model_params=dict(hidden_dim=192, depth=12, num_slices=512),
@@ -91,10 +98,11 @@ def train_transolver(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
+        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 if __name__ == "__main__":
