+++ recipes/aero_cfd_python/train_drivaerml.py
@@ -1,10 +1,12 @@
 #  Copyright © 2026 Emmi AI GmbH. All rights reserved.
 
-from examples.aero_cfd import AhmedMLPreset
+from noether.core.distributed.utils import accelerator_to_device
 from noether.training.runners import HydraRunner
 
-DATASET_ROOT = "/path/to/ahmedml"
-OUTPUT_PATH = "/path/to/outputs/ahmedml"
+from .presets import DrivAerMLPreset
+
+DATASET_ROOT = "/nfs-gpu/research/datasets/drivaerml/preprocessed/subsampled_10x"
+OUTPUT_PATH = "outputs/drivaerml"
 TRAINER_KIND = "noether.training.trainers.WeightedLossTrainer"
 FIELD_WEIGHTS = {
     "surface_pressure": 1.0,
@@ -20,10 +22,9 @@ def train_abupt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains AB-UPT model using AhmedML dataset."""
-    preset = AhmedMLPreset()
+    """Train AB-UPT model using DrivAerML dataset."""
+    preset = DrivAerMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroABUPT",
         model_params=dict(hidden_dim=192, geometry_depth=6, physics_blocks=["perceiver"] + ["shared", "cross"] * 5),
@@ -34,8 +35,9 @@ def train_abupt(
         datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
+        num_workers=0,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_upt(
@@ -43,10 +45,9 @@ def train_upt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains UPT model using AhmedML dataset."""
-    preset = AhmedMLPreset()
+    """Train UPT model using DrivAerML dataset."""
+    preset = DrivAerMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroUPT",
         model_params=dict(hidden_dim=192, num_heads=3, approximator_depth=12),
@@ -58,7 +59,7 @@ def train_upt(
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transformer(
@@ -66,10 +67,9 @@ def train_transformer(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains Transformer model using AhmedML dataset."""
-    preset = AhmedMLPreset()
+    """Train Transformer model using DrivAerML dataset."""
+    preset = DrivAerMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransformer",
         model_params=dict(hidden_dim=192, depth=12),
@@ -81,7 +81,7 @@ def train_transformer(
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transolver(
@@ -89,10 +89,9 @@ def train_transolver(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Trains Transolver model using AhmedML dataset."""
-    preset = AhmedMLPreset()
+    """Train Transolver model using DrivAerML dataset."""
+    preset = DrivAerMLPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransolver",
         model_params=dict(hidden_dim=192, depth=12, num_slices=512),
@@ -104,11 +103,11 @@ def train_transolver(
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 if __name__ == "__main__":
-    train_abupt()
+    train_abupt(accelerator="gpu")
     # train_upt()
     # train_transformer()
     # train_transolver()
