+++ recipes/aero_cfd_python/train_shapenet_car.py
@@ -1,18 +1,15 @@
 #  Copyright © 2026 Emmi AI GmbH. All rights reserved.
 
-from examples.aero_cfd import DrivAerMLPreset
+from noether.core.distributed.utils import accelerator_to_device
 from noether.training.runners import HydraRunner
 
-DATASET_ROOT = "/path/to/drivaerml"
-OUTPUT_PATH = "/path/to/outputs/drivaerml"
+from .presets import ShapeNetCarPreset
+
+DATASET_ROOT = "/path/to/shapenet_car"
+OUTPUT_PATH = "/path/to/outputs/shapenet_car"
+
 TRAINER_KIND = "noether.training.trainers.WeightedLossTrainer"
-FIELD_WEIGHTS = {
-    "surface_pressure": 1.0,
-    "surface_friction": 1.0,
-    "volume_pressure": 1.0,
-    "volume_velocity": 1.0,
-    "volume_vorticity": 1.0,
-}
+FIELD_WEIGHTS = {"surface_pressure": 1.0, "volume_velocity": 1.0}
 
 
 def train_abupt(
@@ -20,10 +17,9 @@ def train_abupt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Train AB-UPT model using DrivAerML dataset."""
-    preset = DrivAerMLPreset()
+    """Trains AB-UPT model using ShapeNetCar dataset."""
+    preset = ShapeNetCarPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroABUPT",
         model_params=dict(hidden_dim=192, geometry_depth=6, physics_blocks=["perceiver"] + ["shared", "cross"] * 5),
@@ -31,11 +27,10 @@ def train_abupt(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
-        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_upt(
@@ -43,10 +38,9 @@ def train_upt(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Train UPT model using DrivAerML dataset."""
-    preset = DrivAerMLPreset()
+    """Trains UPT model using ShapeNetCar dataset."""
+    preset = ShapeNetCarPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroUPT",
         model_params=dict(hidden_dim=192, num_heads=3, approximator_depth=12),
@@ -54,11 +48,10 @@ def train_upt(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
-        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transformer(
@@ -66,10 +59,9 @@ def train_transformer(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Train Transformer model using DrivAerML dataset."""
-    preset = DrivAerMLPreset()
+    """Trains Transformer model using ShapeNetCar dataset."""
+    preset = ShapeNetCarPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransformer",
         model_params=dict(hidden_dim=192, depth=12),
@@ -77,11 +69,10 @@ def train_transformer(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
-        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 def train_transolver(
@@ -89,10 +80,9 @@ def train_transolver(
     dataset_root: str = DATASET_ROOT,
     output_path: str = OUTPUT_PATH,
     accelerator: str = "mps",
-    device: str = "mps",
 ) -> None:
-    """Train Transolver model using DrivAerML dataset."""
-    preset = DrivAerMLPreset()
+    """Trains Transolver model using ShapeNetCar dataset."""
+    preset = ShapeNetCarPreset()
     config = preset.build_config(
         model_kind="noether.modeling.models.aerodynamics.AeroTransolver",
         model_params=dict(hidden_dim=192, depth=12, num_slices=512),
@@ -100,15 +90,14 @@ def train_transolver(
         trainer_params=dict(field_weights=FIELD_WEIGHTS),
         dataset_root=dataset_root,
         output_path=output_path,
-        datasets=["train", "val", "test"],
         max_epochs=2,
         accelerator=accelerator,
     )
-    HydraRunner().main(device=device, config=config)
+    HydraRunner().main(device=accelerator_to_device(accelerator), config=config)
 
 
 if __name__ == "__main__":
-    train_abupt()
+    train_abupt(accelerator="gpu")
     # train_upt()
     # train_transformer()
     # train_transolver()
