--- src/noether/core/schemas/aero/__init__.py
@@ -1,9 +1,0 @@
-#  Copyright © 2026 Emmi AI GmbH. All rights reserved.
-
-from .dataset import AeroDatasetConfig
-from .pipeline import AeroCFDPipelineConfig
-
-__all__ = [
-    "AeroCFDPipelineConfig",
-    "AeroDatasetConfig",
-]
